modified delete category has product

+++ app/Http/Controllers/CategoryController.php
@@ -49,12 +49,12 @@ class CategoryController extends Controller
 
     public function destroy($id)
     {
-        $category = Category::withCount(['child'])->find($id);
-        if ($category->child_count == 0) {
+        $category = Category::withCount(['child', 'product'])->find($id);
+        if ($category->child_count == 0 && $category->product_count == 0) {
             $category->delete();
             return redirect(route('category.index'))->with(['success' => 'Kategori Dihapus!']);
         }
 
-        return redirect(route('category.index'))->with(['error' => 'Kategori Ini Memiliki Anak Kategori!']);
+        return redirect(route('category.index'))->with(['error' => 'Kategori Ini Memiliki Anak Kategori atau Ada Produk Terdaftar Untuk Kategori Ini!']);
     }
 }
